Refactor DTO support helpers
